Add ImpersonateClient to RoverAutoEngine; update RoverReading
Added a new private property `ImpersonateClient` to the `RoverAutoEngine` class, ensuring it is always sent via WebSocket and suppressing IDE warnings for unused private members. Updated the `UltrasonicDistance` property in the `RoverReading` class to change the JSON property name to "ultrasonic_distance" and the data type from `double` to `int`.